set for dbt folder update

+++ clean.sh
@@ -1,13 +1,11 @@
 #!/usr/bin/env bash
 
-rm -rf ./airflow/*
-mkdir ./airflow/dags
-mkdir ./airflow/logs
-mkdir ./airflow/plugins
-
 #docker-compose down --rmi all --volumes 
 docker-compose down
-docker volume rm $(docker volume ls -q)
+
+if [[ -n $(docker volume ls -q) ]]; then
+	docker volume rm $(docker volume ls -q)
+fi
